moved emptyspotdesignator to a separate file
